ref(replay): Refactor eventReplay to split onboarding and eligibility checks from preview render
