+++ static/app/components/events/eventReplay/index.tsx
@@ -1,63 +1,50 @@
-import {Fragment, lazy} from 'react';
-import ReactLazyLoad from 'react-lazyload';
-import styled from '@emotion/styled';
+import {lazy} from 'react';
 
-import {LinkButton} from 'sentry/components/button';
-import NegativeSpaceContainer from 'sentry/components/container/negativeSpaceContainer';
 import ErrorBoundary from 'sentry/components/errorBoundary';
-import {REPLAY_LOADING_HEIGHT} from 'sentry/components/events/eventReplay/constants';
-import {EventReplaySection} from 'sentry/components/events/eventReplay/eventReplaySection';
+import {ReplayClipSection} from 'sentry/components/events/eventReplay/replayClipSection';
 import LazyLoad from 'sentry/components/lazyLoad';
-import LoadingIndicator from 'sentry/components/loadingIndicator';
-import {ReplayGroupContextProvider} from 'sentry/components/replays/replayGroupContext';
 import {replayBackendPlatforms} from 'sentry/data/platformCategories';
-import {t, tct} from 'sentry/locale';
-import {space} from 'sentry/styles/space';
 import type {Event} from 'sentry/types/event';
 import type {Group} from 'sentry/types/group';
-import {getAnalyticsDataForEvent, getAnalyticsDataForGroup} from 'sentry/utils/events';
-import useReplayCountForIssues from 'sentry/utils/replayCount/useReplayCountForIssues';
 import {getReplayIdFromEvent} from 'sentry/utils/replays/getReplayIdFromEvent';
 import {useHasOrganizationSentAnyReplayEvents} from 'sentry/utils/replays/hooks/useReplayOnboarding';
 import {projectCanUpsellReplay} from 'sentry/utils/replays/projectSupportsReplay';
 import useOrganization from 'sentry/utils/useOrganization';
 import useProjectFromSlug from 'sentry/utils/useProjectFromSlug';
-import useRouter from 'sentry/utils/useRouter';
-import {normalizeUrl} from 'sentry/utils/withDomainRequired';
 
-type Props = {
+interface Props {
   event: Event;
   projectSlug: string;
   group?: Group;
-};
-
-export const REPLAY_CLIP_OFFSETS = {
-  durationAfterMs: 5_000,
-  durationBeforeMs: 5_000,
-};
+}
 
 const ReplayOnboardingPanel = lazy(() => import('./replayInlineOnboardingPanel'));
-const ReplayClipPreview = lazy(() => import('./replayClipPreview'));
 
-function EventReplayContent({
-  event,
-  group,
-  replayId,
-}: Props & {replayId: undefined | string}) {
+export default function EventReplay({event, group, projectSlug}: Props) {
   const organization = useOrganization();
-  const {hasOrgSentReplays, fetching} = useHasOrganizationSentAnyReplayEvents();
-  const router = useRouter();
-  const {getReplayCountForIssue} = useReplayCountForIssues();
+  const hasReplaysFeature = organization.features.includes('session-replay');
 
-  if (fetching) {
-    return null;
+  const {hasOrgSentReplays, fetching: fetchingHasSentReplays} =
+    useHasOrganizationSentAnyReplayEvents();
+  const project = useProjectFromSlug({organization, projectSlug});
+
+  const replayId = getReplayIdFromEvent(event);
+
+  if (!hasReplaysFeature || fetchingHasSentReplays) {
+    return false;
   }
 
   const platform = group?.project.platform ?? group?.platform ?? 'other';
   const projectId = group?.project.id ?? event.projectID ?? '';
 
-  // frontend or backend platforms
-  if (!hasOrgSentReplays || (!replayId && replayBackendPlatforms.includes(platform))) {
+  if (replayId) {
+    return <ReplayClipSection event={event} replayId={replayId} group={group} />;
+  }
+
+  if (
+    projectCanUpsellReplay(project) &&
+    (!hasOrgSentReplays || replayBackendPlatforms.includes(platform))
+  ) {
     return (
       <ErrorBoundary mini>
         <LazyLoad
@@ -69,111 +56,5 @@ function EventReplayContent({
     );
   }
 
-  if (!replayId) {
-    return null;
-  }
-
-  const startTimestampMS =
-    'startTimestamp' in event ? event.startTimestamp * 1000 : undefined;
-  const timeOfEvent = event.dateCreated ?? startTimestampMS ?? event.dateReceived;
-  const eventTimestampMs = timeOfEvent ? Math.floor(new Date(timeOfEvent).getTime()) : 0;
-
-  // don't try to construct the url if we don't have a group
-  const eventIdFromRouter = router.params.eventId;
-  const baseUrl = group
-    ? eventIdFromRouter
-      ? normalizeUrl(
-          `/organizations/${organization.slug}/issues/${group.id}/events/${eventIdFromRouter}/`
-        )
-      : normalizeUrl(`/organizations/${organization.slug}/issues/${group.id}/`)
-    : '';
-  const replayUrl = baseUrl ? `${baseUrl}replays/${location.search}/` : '';
-  const seeAllReplaysButton = replayUrl ? (
-    <LinkButton
-      size="sm"
-      to={replayUrl}
-      analyticsEventKey="issue_details.replay_player.clicked_see_all_replays"
-      analyticsEventName="Issue Details: Replay Player Clicked See All Replays"
-    >
-      {t('See All Replays')}
-    </LinkButton>
-  ) : undefined;
-  const replayCount = group ? getReplayCountForIssue(group.id, group.issueCategory) : -1;
-  const overlayContent =
-    seeAllReplaysButton && replayCount && replayCount > 1 ? (
-      <Fragment>
-        <div>
-          {tct('Replay captured [replayCount] users experiencing this issue', {
-            replayCount,
-          })}
-        </div>
-        {seeAllReplaysButton}
-      </Fragment>
-    ) : undefined;
-
-  return (
-    <ReplaySectionMinHeight actions={seeAllReplaysButton}>
-      <ErrorBoundary mini>
-        <ReplayGroupContextProvider groupId={group?.id} eventId={event.id}>
-          <ReactLazyLoad debounce={50} height={448} offset={0} once>
-            <LazyLoad
-              analyticsContext="issue_details"
-              replaySlug={replayId}
-              orgSlug={organization.slug}
-              eventTimestampMs={eventTimestampMs}
-              fullReplayButtonProps={{
-                analyticsEventKey: 'issue_details.open_replay_details_clicked',
-                analyticsEventName: 'Issue Details: Open Replay Details Clicked',
-                analyticsParams: {
-                  ...getAnalyticsDataForEvent(event),
-                  ...getAnalyticsDataForGroup(group),
-                  organization,
-                },
-              }}
-              loadingFallback={
-                <StyledNegativeSpaceContainer testId="replay-loading-placeholder">
-                  <LoadingIndicator />
-                </StyledNegativeSpaceContainer>
-              }
-              LazyComponent={ReplayClipPreview}
-              clipOffsets={REPLAY_CLIP_OFFSETS}
-              overlayContent={overlayContent}
-            />
-          </ReactLazyLoad>
-        </ReplayGroupContextProvider>
-      </ErrorBoundary>
-    </ReplaySectionMinHeight>
-  );
-}
-
-export default function EventReplay({event, group, projectSlug}: Props) {
-  const organization = useOrganization();
-  const hasReplaysFeature = organization.features.includes('session-replay');
-
-  const project = useProjectFromSlug({organization, projectSlug});
-  const canUpsellReplay = projectCanUpsellReplay(project);
-  const replayId = getReplayIdFromEvent(event);
-
-  if (hasReplaysFeature && (replayId || canUpsellReplay)) {
-    return (
-      <EventReplayContent
-        event={event}
-        group={group}
-        projectSlug={projectSlug}
-        replayId={replayId}
-      />
-    );
-  }
-
   return null;
 }
-
-// The min-height here is due to max-height that is set in replayPreview.tsx
-const ReplaySectionMinHeight = styled(EventReplaySection)`
-  min-height: 557px;
-`;
-
-const StyledNegativeSpaceContainer = styled(NegativeSpaceContainer)`
-  height: ${REPLAY_LOADING_HEIGHT}px;
-  margin-bottom: ${space(2)};
-`;
